Clearing socket buffer on seek to prevent resetting timestamp on wrong packet

+++ include/nemesi/rtp.h
@@ -99,7 +99,7 @@ typedef struct {
 
 typedef struct {
     long len;
-    long timestamp;
+    uint32 timestamp;
     double time_sec;
     int fps;
     uint8 pt;
